# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SenaVida API",
        "description": "API REST del backend de SenaVida, plataforma de comunicacion inclusiva para personas sordas en contextos de urgencia medica.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Listar usuarios",
                "description": "Devuelve los usuarios visibles para quien hace la peticion. El super_admin ve todos; el admin_institucional solo los de su propio centro de salud.",
                "operationId": "47249454f1cdb4ad92ff4dd2abad233f",
                "responses": {
                    "200": {
                        "description": "Listado de usuarios"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permiso para listar usuarios"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Registrar un nuevo usuario",
                "description": "Crea un usuario del personal de salud. Solo super_admin o admin_institucional pueden hacerlo, y admin_institucional unicamente dentro de su propio centro de salud. La contrasena se cifra automaticamente con bcrypt.",
                "operationId": "4b986d733289251f0bcc4130dd815af2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "role",
                                    "organizationId",
                                    "healthCenterId",
                                    "unitId"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Enfermera de Prueba"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "enfermera@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "categorizacion",
                                        "enum": [
                                            "super_admin",
                                            "admin_institucional",
                                            "admision",
                                            "categorizacion",
                                            "medico"
                                        ]
                                    },
                                    "organizationId": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "healthCenterId": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "unitId": {
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permiso, o intento de crear fuera del propio centro"
                    },
                    "422": {
                        "description": "Datos invalidos o email duplicado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Ver un usuario",
                "description": "Devuelve el detalle de un usuario. El admin_institucional solo puede ver usuarios de su propio centro.",
                "operationId": "90ca44c038360e66084588a1a5951520",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID del usuario",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del usuario"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permiso para ver este usuario"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Editar un usuario",
                "description": "Actualiza los datos de un usuario. El admin_institucional solo puede editar usuarios de su propio centro.",
                "operationId": "e734c66a7d375609edee2154bc5f7b47",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID del usuario",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Enfermera Actualizada"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "nueva@test.com"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "super_admin",
                                            "admin_institucional",
                                            "admision",
                                            "categorizacion",
                                            "medico"
                                        ]
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "nuevaClave123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "nuevaClave123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permiso para editar este usuario"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    },
                    "422": {
                        "description": "Datos invalidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Desactivar un usuario",
                "description": "Marca al usuario como inactivo (isActive=false). No elimina el registro de la base de datos.",
                "operationId": "9daa558fb80f5a3e2b490a3a5b05a716",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID del usuario",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario desactivado"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permiso, o intento de autodesactivacion"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token Bearer obtenido en /auth/login",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}